Redesign chapters dialog in reader
Also, add a progress bar to show completion of a specific chapter
Added two new fonts
Code cleanup & some optimizations
