+++ app/src/main/java/com/starry/myne/ui/screens/reader/activities/ReaderActivity.kt
@@ -23,83 +23,11 @@ import android.os.Handler
 import android.os.Looper
 import android.view.WindowManager
 import android.widget.Toast
-import androidx.activity.compose.BackHandler
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.Crossfade
-import androidx.compose.animation.animateContentSize
-import androidx.compose.animation.expandVertically
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
-import androidx.compose.animation.shrinkVertically
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.displayCutoutPadding
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.selection.selectable
-import androidx.compose.foundation.selection.selectableGroup
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.ButtonDefaults
-import androidx.compose.material.CircularProgressIndicator
 import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.OutlinedButton
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Menu
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.Divider
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.RadioButton
-import androidx.compose.material3.RadioButtonDefaults
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.SnackbarHost
-import androidx.compose.material3.SnackbarHostState
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults
-import androidx.compose.material3.surfaceColorAtElevation
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.SideEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.semantics.Role
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.compose.ui.viewinterop.AndroidView
 import androidx.core.view.WindowCompat
 import androidx.core.view.WindowInsetsCompat
@@ -107,22 +35,19 @@ import androidx.core.view.WindowInsetsControllerCompat
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import com.starry.myne.R
 import com.starry.myne.databinding.ActivityReaderBinding
 import com.starry.myne.ui.screens.reader.adapters.ReaderClickListener
 import com.starry.myne.ui.screens.reader.adapters.ReaderRVAdapter
-import com.starry.myne.ui.screens.reader.viewmodels.ReaderFont
+import com.starry.myne.ui.screens.reader.composables.ReaderScreen
+import com.starry.myne.ui.screens.reader.composables.TransparentSystemBars
 import com.starry.myne.ui.screens.reader.viewmodels.ReaderViewModel
 import com.starry.myne.ui.screens.settings.viewmodels.SettingsViewModel
 import com.starry.myne.ui.screens.settings.viewmodels.ThemeMode
 import com.starry.myne.ui.theme.MyneTheme
-import com.starry.myne.ui.theme.figeronaFont
 import com.starry.myne.utils.toToast
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.launch
 import java.io.FileInputStream
 import kotlin.properties.Delegates
 
@@ -234,7 +159,7 @@ class ReaderActivity : AppCompatActivity(), ReaderClickListener {
             override fun onScrollStateChanged(recyclerView: RecyclerView, newState: Int) {
                 super.onScrollStateChanged(recyclerView, newState)
                 // fetch last visible chapter position and offset.
-                val progressChIndex = layoutManager.findLastVisibleItemPosition()
+                val progressChIndex = layoutManager.findFirstVisibleItemPosition()
                 val progressChOffset = layoutManager.findViewByPosition(progressChIndex)!!.top
                 // Emit currently visible chapter.
                 visibleChapterFlow.value = progressChIndex
@@ -248,19 +173,26 @@ class ReaderActivity : AppCompatActivity(), ReaderClickListener {
                     )
                 }
             }
+
+            override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
+                super.onScrolled(recyclerView, dx, dy)
+                val firstVisiblePosition = layoutManager.findFirstVisibleItemPosition()
+                // Calculate the scroll percentage for the currently visible chapter
+                // and store its value in view model.
+                viewModel.setItemScrolledPercent(
+                    calculateItemScrollPercentage(recyclerView, firstVisiblePosition)
+                )
+            }
         })
 
         // Set UI contents.
         setContent {
             MyneTheme(settingsViewModel = settingsViewModel) {
-                TransparentSystemBars()
-                val textSizeValue = remember { mutableIntStateOf(viewModel.getFontSize()) }
-                val readerFontFamily = remember { mutableStateOf(viewModel.getFontFamily()) }
+                TransparentSystemBars(settingsViewModel = settingsViewModel)
                 ReaderScreen(
                     viewModel = viewModel,
-                    textSizeValue = textSizeValue,
-                    readerFontFamily = readerFontFamily,
                     recyclerViewManager = layoutManager,
+                    visibleChapterFlow = visibleChapterFlow,
                     readerContent = { AndroidView(factory = { binding.root }) }
                 )
             }
@@ -298,448 +230,38 @@ class ReaderActivity : AppCompatActivity(), ReaderClickListener {
         }
     }
 
-    @Composable
-    private fun ReaderScreen(
-        viewModel: ReaderViewModel,
-        textSizeValue: MutableState<Int>,
-        readerFontFamily: MutableState<ReaderFont>,
-        recyclerViewManager: LinearLayoutManager,
-        readerContent: @Composable (paddingValues: PaddingValues) -> Unit
-    ) {
-        // Hide reader menu on back press.
-        BackHandler(viewModel.state.showReaderMenu) {
-            viewModel.hideReaderInfo()
-        }
-
-        // Font style chooser dialog.
-        val showFontDialog = remember { mutableStateOf(false) }
-        FontChooserDialog(showFontDialog, viewModel, readerFontFamily)
-
-        // Chapter chooser dialog.
-        val showChaptersDialog = remember { mutableStateOf(false) }
-        ChaptersDialog(recyclerViewManager, showChaptersDialog)
-
-        val snackBarHostState = remember { SnackbarHostState() }
-        val visibleChapterIdx by visibleChapterFlow.collectAsState()
-        val chapter = viewModel.state.epubBook?.chapters?.getOrNull(visibleChapterIdx)
-
-        Scaffold(
-            snackbarHost = { SnackbarHost(snackBarHostState) },
-            topBar = {
-                AnimatedVisibility(
-                    visible = viewModel.state.showReaderMenu,
-                    enter = expandVertically(initialHeight = { 0 }, expandFrom = Alignment.Top)
-                            + fadeIn(),
-                    exit = shrinkVertically(targetHeight = { 0 }, shrinkTowards = Alignment.Top)
-                            + fadeOut(),
-                ) {
-                    Surface(color = MaterialTheme.colorScheme.surfaceColorAtElevation(2.dp)) {
-                        Column(modifier = Modifier.displayCutoutPadding()) {
-                            TopAppBar(
-                                colors = TopAppBarDefaults.topAppBarColors(
-                                    containerColor = MaterialTheme.colorScheme.surfaceColorAtElevation(
-                                        2.dp
-                                    ),
-                                    scrolledContainerColor = MaterialTheme.colorScheme.surfaceColorAtElevation(
-                                        2.dp
-                                    ),
-                                ),
-                                title = {
-                                    viewModel.state.epubBook?.let {
-                                        Text(
-                                            text = it.title,
-                                            maxLines = 1,
-                                            overflow = TextOverflow.Ellipsis,
-                                            modifier = Modifier.animateContentSize(),
-                                            color = MaterialTheme.colorScheme.onSurface
-                                        )
-                                    }
-                                },
-                                actions = {
-                                    IconButton(onClick = { showChaptersDialog.value = true }) {
-                                        Icon(
-                                            Icons.Filled.Menu, null,
-                                            tint = MaterialTheme.colorScheme.onSurface
-                                        )
-                                    }
-                                }
-                            )
-                            Column(
-                                modifier = Modifier
-                                    .padding(bottom = 8.dp)
-                                    .padding(horizontal = 16.dp),
-                            ) {
-                                chapter?.title?.let {
-                                    Text(
-                                        text = it,
-                                        maxLines = 2,
-                                        overflow = TextOverflow.Ellipsis,
-                                        color = MaterialTheme.colorScheme.onSurface
-                                    )
-                                }
-                            }
-                            Divider()
-                        }
-                    }
-                }
-            },
-            bottomBar = {
-                AnimatedVisibility(
-                    visible = viewModel.state.showReaderMenu,
-                    enter = expandVertically(initialHeight = { 0 }) + fadeIn(),
-                    exit = shrinkVertically(targetHeight = { 0 }) + fadeOut(),
-                ) {
-                    BottomSheetContents(
-                        viewModel = viewModel,
-                        textSizeValue = textSizeValue,
-                        readerFontFamily = readerFontFamily,
-                        showFontDialog = showFontDialog,
-                        snackBarHostState = snackBarHostState
-                    )
-                }
-            },
-            content = { paddingValues ->
-                Crossfade(
-                    targetState = viewModel.state.isLoading,
-                    label = "reader content loading cross fade"
-                ) { loading ->
-                    if (loading) {
-                        Box(
-                            modifier = Modifier
-                                .fillMaxSize()
-                                .padding(bottom = 65.dp),
-                            contentAlignment = Alignment.Center
-                        ) {
-                            CircularProgressIndicator(color = MaterialTheme.colorScheme.primary)
-                        }
-                    } else {
-                        readerContent(paddingValues)
-                    }
-                }
+    /**
+     * Calculate the scroll percentage for a specific item in a RecyclerView.
+     *
+     * @param recyclerView The RecyclerView containing the item
+     * @param itemPosition The position of the item in the RecyclerView
+     * @return The scroll percentage for the item, or -1 if the item is not visible
+     */
+    fun calculateItemScrollPercentage(recyclerView: RecyclerView, itemPosition: Int): Float {
+        val layoutManager = recyclerView.layoutManager!!
+        // Find the view corresponding to the item position
+        val itemView = layoutManager.findViewByPosition(itemPosition) ?: return -1f
+        // Calculate the vertical offset of the item relative to the RecyclerView
+        val itemTop = itemView.top.toFloat()
+        val itemBottom = itemView.bottom.toFloat()
+        // Calculate the visible height of the RecyclerView
+        val recyclerViewHeight = recyclerView.height.toFloat()
+        // Calculate the height of the item
+        val itemHeight = itemView.height.toFloat()
+        // Calculate the scroll percentage for the item
+        return if (itemTop >= recyclerViewHeight || itemBottom <= 0f) {
+            1f // Item is completely scrolled out of view
+        } else {
+            // Calculate the visible portion of the item
+            val visiblePortion = if (itemTop < 0) {
+                itemBottom
+            } else {
+                recyclerViewHeight - itemTop
             }
-        )
-    }
-
-    @Composable
-    private fun FontChooserDialog(
-        showFontDialog: MutableState<Boolean>,
-        viewModel: ReaderViewModel,
-        readerFontFamily: MutableState<ReaderFont>
-    ) {
-        val radioOptions = ReaderFont.getAllFonts().map { it.name }
-        val (selectedOption, onOptionSelected) = remember {
-            mutableStateOf(viewModel.getFontFamily())
-        }
-
-        if (showFontDialog.value) {
-            AlertDialog(onDismissRequest = {
-                showFontDialog.value = false
-            }, title = {
-                Text(
-                    text = stringResource(id = R.string.reader_font_style_chooer),
-                    color = MaterialTheme.colorScheme.onSurface,
-                )
-            }, text = {
-                Column(
-                    modifier = Modifier.selectableGroup(),
-                    verticalArrangement = Arrangement.Center,
-                ) {
-                    radioOptions.forEach { text ->
-                        Row(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .height(46.dp)
-                                .selectable(
-                                    selected = (text == selectedOption.name),
-                                    onClick = { onOptionSelected(ReaderFont.getFontByName(text)) },
-                                    role = Role.RadioButton,
-                                ),
-                            verticalAlignment = Alignment.CenterVertically,
-                        ) {
-                            RadioButton(
-                                selected = (text == selectedOption.name),
-                                onClick = null,
-                                colors = RadioButtonDefaults.colors(
-                                    selectedColor = MaterialTheme.colorScheme.primary,
-                                    unselectedColor = MaterialTheme.colorScheme.inversePrimary,
-                                    disabledSelectedColor = Color.Black,
-                                    disabledUnselectedColor = Color.Black
-                                ),
-                            )
-                            Text(
-                                text = text,
-                                modifier = Modifier.padding(start = 16.dp),
-                                color = MaterialTheme.colorScheme.onSurface,
-                                fontFamily = figeronaFont
-                            )
-                        }
-                    }
-                }
-            }, confirmButton = {
-                TextButton(onClick = {
-                    showFontDialog.value = false
-                    viewModel.setFontFamily(selectedOption)
-                    readerFontFamily.value = selectedOption
-                }) {
-                    Text(stringResource(id = R.string.theme_dialog_apply_button))
-                }
-            }, dismissButton = {
-                TextButton(onClick = {
-                    showFontDialog.value = false
-                }) {
-                    Text(stringResource(id = R.string.cancel))
-                }
-            })
-        }
-    }
-
-    @Composable
-    private fun ChaptersDialog(
-        recyclerViewManager: LinearLayoutManager,
-        showChaptersDialog: MutableState<Boolean>
-    ) {
-        if (showChaptersDialog.value) {
-            AlertDialog(onDismissRequest = {
-                showChaptersDialog.value = false
-            }, content = {
-                Column(
-                    modifier = Modifier.background(MaterialTheme.colorScheme.surface)
-                ) {
-                    LazyColumn(content = {
-                        viewModel.state.epubBook?.chapters?.size?.let {
-                            items(it) { index ->
-                                val chapter = viewModel.state.epubBook!!.chapters[index]
-                                TextButton(
-                                    onClick = {
-                                        recyclerViewManager.scrollToPositionWithOffset(index, 0)
-                                        showChaptersDialog.value = false
-                                        viewModel.hideReaderInfo()
-                                    },
-                                    modifier = Modifier
-                                        .fillMaxWidth()
-                                        .padding(4.dp)
-                                ) {
-                                    Text(
-                                        text = chapter.title,
-                                        modifier = Modifier.fillMaxWidth(),
-                                        textAlign = TextAlign.Start
-                                    )
-                                }
-                            }
-                        }
-                    })
-                }
-            })
-        }
-    }
-
-    @Composable
-    private fun BottomSheetContents(
-        viewModel: ReaderViewModel,
-        textSizeValue: MutableState<Int>,
-        readerFontFamily: MutableState<ReaderFont>,
-        showFontDialog: MutableState<Boolean>,
-        snackBarHostState: SnackbarHostState
-    ) {
-        val coroutineScope = rememberCoroutineScope()
-
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(MaterialTheme.colorScheme.surfaceColorAtElevation(2.dp))
-                .padding(vertical = 24.dp, horizontal = 16.dp)
-        ) {
-            TextScaleControls(
-                textSizeValue = textSizeValue,
-                viewModel = viewModel,
-                coroutineScope = coroutineScope,
-                snackBarHostState = snackBarHostState
-            )
-
-            Spacer(modifier = Modifier.height(14.dp))
-            Divider()
-
-            FontSelectionButton(
-                readerFontFamily = readerFontFamily,
-                showFontDialog = showFontDialog
-            )
-        }
-    }
-
-    @Composable
-    private fun TextScaleControls(
-        textSizeValue: MutableState<Int>,
-        viewModel: ReaderViewModel,
-        coroutineScope: CoroutineScope,
-        snackBarHostState: SnackbarHostState
-    ) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.Center
-        ) {
-            ReaderTextScaleButton(
-                buttonType = TextScaleButtonType.DECREASE,
-                textSizeState = textSizeValue,
-                coroutineScope = coroutineScope,
-                snackBarHostState = snackBarHostState,
-                onTextSizeValueChanged = { viewModel.setFontSize(it) }
-            )
-
-            Spacer(modifier = Modifier.width(14.dp))
-
-            Box(
-                modifier = Modifier
-                    .size(100.dp, 45.dp)
-                    .background(MaterialTheme.colorScheme.surfaceColorAtElevation(2.dp))
-                    .border(
-                        width = 1.dp,
-                        color = MaterialTheme.colorScheme.onSurface,
-                        shape = RoundedCornerShape(6.dp)
-                    )
-                    .clip(RoundedCornerShape(6.dp)),
-                contentAlignment = Alignment.Center
-            ) {
-                Text(
-                    text = textSizeValue.value.toString(),
-                    fontFamily = figeronaFont,
-                    fontSize = 16.sp,
-                    color = MaterialTheme.colorScheme.onSurface,
-                    modifier = Modifier.padding(start = 2.dp, bottom = 1.dp)
-                )
-            }
-
-            Spacer(modifier = Modifier.width(14.dp))
-
-            ReaderTextScaleButton(
-                buttonType = TextScaleButtonType.INCREASE,
-                textSizeState = textSizeValue,
-                coroutineScope = coroutineScope,
-                snackBarHostState = snackBarHostState,
-                onTextSizeValueChanged = { viewModel.setFontSize(it) }
-            )
-        }
-    }
-
-    @Composable
-    private fun FontSelectionButton(
-        readerFontFamily: MutableState<ReaderFont>,
-        showFontDialog: MutableState<Boolean>
-    ) {
-        OutlinedButton(
-            onClick = { showFontDialog.value = true },
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 14.dp),
-            shape = RoundedCornerShape(16.dp),
-            border = BorderStroke(1.dp, MaterialTheme.colorScheme.onSurface),
-            colors = ButtonDefaults.buttonColors(
-                backgroundColor = MaterialTheme.colorScheme.surfaceColorAtElevation(2.dp)
-            )
-        ) {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                Icon(
-                    imageVector = ImageVector.vectorResource(id = R.drawable.ic_reader_font),
-                    contentDescription = null,
-                    tint = MaterialTheme.colorScheme.onSurface
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(
-                    text = readerFontFamily.value.name,
-                    color = MaterialTheme.colorScheme.onSurface
-                )
-            }
-        }
-    }
-
-    @Composable
-    private fun ReaderTextScaleButton(
-        buttonType: TextScaleButtonType,
-        textSizeState: MutableState<Int>,
-        coroutineScope: CoroutineScope,
-        snackBarHostState: SnackbarHostState,
-        onTextSizeValueChanged: (newValue: Int) -> Unit
-    ) {
-        val context = LocalContext.current
-        val iconRes: Int
-        val adjustment: Int
-
-        when (buttonType) {
-            TextScaleButtonType.DECREASE -> {
-                iconRes = R.drawable.ic_reader_text_minus
-                adjustment = -10
-            }
-
-            TextScaleButtonType.INCREASE -> {
-                iconRes = R.drawable.ic_reader_text_plus
-                adjustment = 10
-            }
-        }
-
-        val callback: () -> Unit = {
-            val newSize = textSizeState.value + adjustment
-            when {
-                newSize < 50 -> {
-                    coroutineScope.launch {
-                        snackBarHostState.showSnackbar(
-                            context.getString(R.string.reader_min_font_size_reached),
-                            null
-                        )
-                    }
-                }
-
-                newSize > 200 -> {
-                    coroutineScope.launch {
-                        snackBarHostState.showSnackbar(
-                            context.getString(R.string.reader_max_font_size_reached),
-                            null
-                        )
-                    }
-                }
-
-                else -> {
-                    coroutineScope.launch {
-                        val adjustedSize = textSizeState.value + adjustment
-                        textSizeState.value = adjustedSize
-                        onTextSizeValueChanged(adjustedSize)
-                    }
-                }
-            }
-        }
-
-        Box(
-            modifier = Modifier
-                .size(100.dp, 45.dp)
-                .background(MaterialTheme.colorScheme.surfaceColorAtElevation(2.dp))
-                .border(
-                    1.dp,
-                    MaterialTheme.colorScheme.onSurface,
-                    shape = RoundedCornerShape(6.dp)
-                )
-                .clip(RoundedCornerShape(6.dp))
-                .clickable { callback() },
-            contentAlignment = Alignment.Center
-        ) {
-            Icon(
-                imageVector = ImageVector.vectorResource(id = iconRes),
-                contentDescription = stringResource(id = R.string.back_button_desc),
-                tint = MaterialTheme.colorScheme.onSurface,
-                modifier = Modifier.padding(14.dp)
-            )
+            // Calculate the scroll percentage based on the visible portion
+            ((1 - visiblePortion / itemHeight)).coerceIn(0f, 1f)
         }
     }
 
 
-    @Composable
-    private fun TransparentSystemBars(alpha: Float = 0f) {
-        val systemUiController = rememberSystemUiController()
-        val useDarkIcons = settingsViewModel.getCurrentTheme() == ThemeMode.Light
-        val baseColor = MaterialTheme.colorScheme.primary
-        val color = remember(alpha, baseColor) { baseColor.copy(alpha = alpha) }
-        SideEffect {
-            systemUiController.setSystemBarsColor(
-                color = color,
-                darkIcons = useDarkIcons
-            )
-        }
-    }
 }
