+++ app/src/main/res/values-ru/strings.xml
@@ -62,6 +62,7 @@
     <string name="reader_max_font_size_reached">Извините, достигнут максимальный размер шрифта!</string>
     <string name="reader_min_font_size_reached">Извините, достигнут минимальный размер шрифта!</string>
     <string name="reader_font_style_chooer">Стиль шрифта</string>
+    <string name="reader_chapter_list_title">Главы</string>
 
     <!-- Settings screen. -->
     <string name="settings_header">Настройки</string>
